Always leave at least one item in S3 bucket in clean_s3 and handle case where bucket is empty or missing (#13)
* Revert "Revert "Always leave at least one item in S3 bucket in clean_s3""

* Fix case where bucket is empty

* Add logging around getting number of items in bucket

* Fix error handling syntax

* Add more logging

Breaking in some projects for some reason

* Add still more logging

* Log level: ludicrous

* Try another way of incrementing

* Remove some extraneous logging

Determined the previous incrementing didn't work b/c of set -e

* Trim down the logging

+++ deploy.sh
@@ -140,6 +140,11 @@ then
 
     for suffix in deploy pull-request
     do
+        # track number of items in the bucket to ensure we don't delete everything, which would break the _deploy servlet
+        item_count=0
+        echo "Getting number of items in $DEPLOY_BUCKET with prefix $cleanup_prefix$suffix/..."
+        number_of_items=`aws s3api list-objects --bucket $DEPLOY_BUCKET --prefix $cleanup_prefix$suffix/ --output=json --query="length(Contents[])"` || number_of_items=0
+        echo "$number_of_items"
         aws s3api list-objects --bucket $DEPLOY_BUCKET --prefix $cleanup_prefix$suffix/ --output=text | \
         while read -r line
         do
@@ -148,14 +153,19 @@ then
             then
                 continue
             fi
+            item_count=$((item_count+1))
             last_modified_ts=`date -d"$last_modified" +%s`
             filename=`echo "$line" | awk -F'\t' '{print $3}'`
+            echo "File # $item_count: $filename. Last modified: $last_modified_ts"
             if [[ $last_modified_ts -lt $older_than_ts ]]
             then
-                if [[ $filename != "" ]]
+                if [[ $filename != "" && "$item_count" -ne "$number_of_items" ]]
                 then
                     echo "s3://$DEPLOY_BUCKET/$filename is older than $PURGE_OLDER_THAN_DAYS days ($last_modified). Deleting."
                     aws s3 rm "s3://$DEPLOY_BUCKET/$filename"
+                elif [[ $filename != "" ]]
+                then
+                    echo "Skipping delete on s3://$DEPLOY_BUCKET/$filename to ensure at least one file in $DEPLOY_BUCKET."
                 fi
             fi
         done
